Finalised the driver stepper motor code. The code now is finished and can be used in the final stages of the robot. Only minor adjustments might be nececary for the system to function optimally, apart from that it is all ready!

+++ Code/PrimitiveStepper.h
@@ -20,7 +20,7 @@ protected:
 	//Total steps the motor has made. Useful for absolute positioning.
 	volatile int32_t currentSteps;
 	//Steps the motor still has to make.
-	volatile int32_t stepsToGo;
+	volatile float stepsToGo;
 
 	//Steps per ISR Call.
 	volatile float stepSpeed;
